requirements.txt
google-genai
groq
python-dotenv